### 2021-01-14 #2953 @whedon
[PRE REVIEW]: splithalf: robust estimates of split half reliability
**Submitting author:** @sdparsons (<a href="http://orcid.org/0000-0002-7048-4093">Sam Parsons</a>)
**Repository:** <a href="https://github.com/sdparsons/splithalf/" target ="_blank">https://github.com/sdparsons/splithalf/</a>
**Version:** 0.7.2
