+++ inst/book.bib
@@ -61,7 +61,7 @@
 @misc{parsons_kruijt_fox_2019,
  title={Psychological Science needs a standard practice of reporting the reliability of cognitive behavioural measurements},
  url={https://doi.org/10.1177/2515245919879695},
- DOI={https://doi.org/10.1177/2515245919879695},
+ DOI={doi.org/10.1177/2515245919879695},
  publisher={Advances in Methods and Practices in Psychological Science},
  author={Sam Parsons and Anne-Wil Kruijt and Elaine Fox},
  year={2019},
