+++ README.md
@@ -9,14 +9,14 @@ Version](http://www.r-pkg.org/badges/version/splithalf)](https://cran.r-project.
 
 # splithalf: robust estimates of split half reliability
 
-The `R` package *splithalf* provides tools to estimate the internal
+The `R` package **splithalf** provides tools to estimate the internal
 consistency reliability of cognitive measures. In particular, the tools
 were developed for application to tasks that use difference scores as
 the main outcome measure, for instance the Stroop score or dot-probe
 attention bias index (average RT in incongruent trials minus average RT
 in congruent trials).
 
-The methods in *splithalf* are built around split half reliability
+The methods in **splithalf** are built around split half reliability
 estimation. To increase the robustness of these estimates, the package
 implements a permutation approach that takes a large number of random
 (without replacement) split halves of the data. For each permutation the
@@ -34,7 +34,7 @@ questions of individual differences researchers need to have some
 psychometric information for the outcome measures. Recently, it was
 proposed that psychological science should set a standard expectation
 for the reporting of reliability information for cognitive and
-behavioural measures (2019). *splithalf* was developed to support this
+behavioural measures (2019). **splithalf** was developed to support this
 proposition by providing a tool to easily extract internal consistency
 reliability estimates from behavioural measures. \#\# Installation
 
@@ -51,15 +51,15 @@ The current developmental version can be installed from Github with:
 devtools::install_github("sdparsons/splithalf")
 ```
 
-*splithalf* requires the *tidyr* (Wickham and Henry 2019) and *dplyr*
-(Wickham et al. 2018) packages for data handling within the functions.
-The *robustbase* package is used to extract median scores when
-applicable. The computationally heavy tasks (extracting many random half
-samples of the data) are written in `c++` via the `R` package *Rcpp*
-(Eddelbuettel et al. 2018). Figures use the *ggplot* package (Wickham
-2016), raincloud plots use code adapted from Allen et al. (Allen et al.
-2019), and the *patchwork* package (Pedersen 2019) is used for plotting
-the multiverse analyses.
+**splithalf** requires the **tidyr** (Wickham and Henry 2019) and
+**dplyr** (Wickham et al. 2018) packages for data handling within the
+functions. The **robustbase** package is used to extract median scores
+when applicable. The computationally heavy tasks (extracting many random
+half samples of the data) are written in `c++` via the `R` package
+**Rcpp** (Eddelbuettel et al. 2018). Figures use the **ggplot** package
+(Wickham 2016), raincloud plots use code adapted from Allen et
+al. (Allen et al. 2019), and the **patchwork** package (Pedersen 2019)
+is used for plotting the multiverse analyses.
 
 ### Citing the package
 
@@ -195,7 +195,7 @@ need, and are aimed to make the *splithalf* function easy to use.
 
 Are you interested in response times, or accuracy rates?
 
-Knowing this, you can set *outcome* = “RT”, or *outcome* = “accuracy”
+Knowing this, you can set `outcome = "RT"`, or `outcome = "accuracy"`
 
 2.  **How is your outcome score calculated?**
 
@@ -203,8 +203,8 @@ Say that your response time based task has two trial types;
 “incongruent” and “congruent”. When you analyse your data will you
 use the average RT in each trial type, or will you create a difference
 score (or bias) by e.g. subtracting the average RT in congruent trials
-from the average RT in incongruent trials. The first I call “average”
-and the second I call “difference”.
[…]
