### 2021-01-14 #2953 @sdparsons
I have corrected the DOI issue in the .bib file and checked the generated pdf.
