### 2021-01-14 #2953 @Kevin-Mattheus-Moerman
@sdparsons I've flagged this submission for additional scope review by our editorial board. In particular we will review if it is considered to have [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This is standard for submissions of this size (the core functionality appears to be captured by around 1366 lines of code).  This review will take about a week.
