### 2021-02-02 #2953 @sdparsons
Great! I've gathered the details for five potential reviewers and also noted relevant expertise. I hope this helps, and I can think of more people if needed too.

James Bartlett (bartle16@uni.coventry.ac.uk) – has used the package before (I’m not sure if this constitutes a COI for JOSS – though I will note that we have worked together on a separate project before also) and also done work with task reliability.
Shelly Cooper (shelly.cooper@wustl.edu) – has worked on task reliability and implementing splithalf-reliability estimation procedures
Jean-Paul Snijder (jean-paul.snijder@cgu.edu) – has worked on reliability issues and used similar methods as implemented in the package
Donald Williams (drrwilliams@ucdavis.edu) - Recently published in JOSS and has a good record in psychometrics
Julia Haaf (j.m.haaf@uva.nl) strong psychometrician and R expertise
