### 2021-02-15 #3041 @whedon
[REVIEW]: splithalf: robust estimates of split half reliability
**Submitting author:** @sdparsons (<a href="http://orcid.org/0000-0002-7048-4093">Sam Parsons</a>)
**Repository:** <a href="https://github.com/sdparsons/splithalf/" target ="_blank">https://github.com/sdparsons/splithalf/</a>
**Version:** v0.8.1
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.11956746.v5" target="_blank">10.6084/m9.figshare.11956746.v5</a>
