### 2021-02-23 #3041 @jpsnijder
Dear @oliviaguest,

here are some additional notes and comments on the current review. If items are not listed here, then they were checked off without any issues or warranting a comment.

Overall, the package is in good shape and I can't foresee any issues with widespread use of it.
minor: https://github.com/sdparsons/splithalf states that the current version = 0.7.1; provided code installs 0.7.2
(I don't see why I would open an issue for this @sdparsons)

**General checks**
 All ✅
**Functionality**
 All ✅
Additionally;
- installs fine on current R/Rstudio versions, as well as on an older version (R3.6).
- ran ***all*** functions on private data; no issues and run as expected.
--------------------------------

@oliviaguest are the items under the following two headings _swapped_ by any chance? Under **documentation** is refers to the "paper". I will update these when I hear back from you.
**Documentation**
❓  A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
**Software paper**
❓  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
