### 2021-03-01 71e25df
Add SD option for type of "average"
Add an option to calculate the standard deviations instead of mean or median for the
"average" argument. Uses plyr::colwise() for the column-wise SD calculation.
