### 2021-03-03 #3041 @oliviaguest
@jpsnijder thanks for your review so far!

I suggest you submit a PR for the clarification you require — probably it's to this file, if I understand your requirements: https://github.com/openjournals/joss/blob/master/app/views/shared/review_body.text.erb
