### 2021-03-04 #3041 @jpsnijder
Hi @oliviaguest, I am not sure how that works exactly. But I have submitted an issue with the journal as well.
Just to clarify: Under the Documentation section of the checklist there is the statement of need item that refers to "the paper" (i.e., _A statement of need: Does the paper have a section titled_). Does "the paper" here refer to the software paper or the documentation? If it refers to the software paper then it seems like the items are under their wrong respective sections (or; they are swapped). Before I complete the review, I figured I get some clarification on this.
