### 2021-03-04 sdparsons/splithalf#9 @Nathaniel-Haines
Validation of estimates
Hi @sdparsons, I am wondering if you have run any validation analyses to ensure that the estimates you are getting for splithalf results are what you expect? I checked the psych methods paper that you reference in the documentation, but was not able to find anything specifically on validating the software.

For validation, I am referring something like simulating with a known reliability and using `splithalf::splithalf` to recover it, or alternatively comparing the output of `splithalf::splithalf` to other packages. For the latter case, I wrote up some code to compare `splithalf::splithalf` to `psych::splitHalf`, which offers similar functionality. As you will see from the `R` code and corresponding distributions of estimates below, they do return similar results for one simple example, although they are not identical. I wonder if it would be worth doing a more comprehensive comparison for validation (e.g., as in the splithalfr package described here: https://github.com/tpronk/splithalfr_simulation).

Here is the code I used to run a simple validation check, along with a plot of the distribution of estimates at the end:

`# compare with other package results
library(splithalf)
library(psych)
library(tidyr)
library(dplyr)
library(patchwork)

# Simulate data
set.seed(43201)
n_participants <- 60 ## sample size
n_trials <- 80
n_blocks <- 2
sim_data <- data.frame(participant_number = rep(1:n_participants, each = n_blocks * n_trials),
                       trial_number = rep(1:n_trials, times = n_blocks * n_participants),
                       block_name = rep(c("A","B"), each = n_trials, length.out = n_participants * n_trials * n_blocks),
                       trial_type = rep(c("congruent","incongruent"), length.out = n_participants * n_trials * n_blocks),
                       RT = rnorm(n_participants * n_trials * n_blocks, 500, 200),
                       ACC = 1)

# create wide data for psych::SplitHalf
test_dat <- sim_data %>%
  filter(trial_type=="congruent" & block_name=="A") %>%
  select(participant_number, trial_number, RT) %>%
  pivot_wider(names_from = trial_number,
              values_from = RT,
              names_prefix = "T") %>%
  select(-participant_number)

# splithalf::splithalf output
splithalf_res <- sim_data %>%
  filter(trial_type=="congruent" & block_name=="A") %>%
  splithalf(data = .,
            outcome = "RT",
            score = "average",
            conditionlist = c("A"),
            halftype = "random",
            permutations = 5000,
            var.RT = "RT",
            var.condition = "block_name",
            var.participant = "participant_number",
            var.compare = "trial_type",
            average = "mean",
            plot = TRUE)

# psych::SplitHalf output
psych_res <- splitHalf(test_dat, n.sample = 5000, check.keys = F, raw = T)
psych_res

# compare psych to splithalf reseults
p1 <- qplot(psych_res$raw,
            main = "psych\nraw values",
            xlim = c(-.6, .6),
            ylim = c(0, 900),
            xlab = "")
p2 <- qplot(splithalf_res$estimates$spearmanbrown,
            main = "splithalf\nSpearman-Brown",
            xlim = c(-.6, .6),
            ylim = c(0, 900),
            xlab = "Split-Half Reliability")
p3 <- qplot(splithalf_res$estimates$splithalf,
            main = "splithalf\nuncorrected",
            xlim = c(-.6, .6),
            ylim = c(0, 900),
            xlab = "")

p1 | p2 | p3`

![plot_zoom_png](https://user-images.githubusercontent.com/25613611/110004906-ecf89080-7ce5-11eb-9b8f-3c9a6f66f1c1.png)
