### 2021-03-04 #3041 @Nathaniel-Haines
Hi @oliviaguest, after passing through everything, I agree with @jpsnijder that things look good overall! I do have a few concerns that I think are worth addressing. The minor ones are those in the comment in this thread directly above this message (here: https://github.com/openjournals/joss-reviews/issues/3041#issuecomment-790718529). I have a few other less minor concerns that I submitted as issues:

1) The first is listed here: https://github.com/sdparsons/splithalf/issues/8, which details an error I get when using the plotting function for one of the main analyses. Note that the results look fine, but the plot is not generated correctly.

2) The second is listed here: https://github.com/sdparsons/splithalf/issues/9, which is a concern regarding validation/whether the package performs as expected. In the simple example I ran, it does give results consistent with, but not identical to, a different package designed to estimate split-half reliability. That said, I think it would be good for users to see that the main functionality has been validated either with some simple simulations or with more comprehensive comparison against other similar software results. I think this could also help contextualize how the package is different from other common packages in the field, which was not immediately apparent to me from what is described in the documentation until I compared them in more detail.

For these reason, I have left the "Functionality", "Automated tests", and "State of the field" checks unticked for now until we get a chance to discuss in more detail. Note that I have also left the "References" check unticked due to (3) in my aforementioned minor comment.
