### 2021-03-05 sdparsons/splithalf#9 @sdparsons
Thanks @Nathaniel-Haines !
I can certainly look further into simulating to validate the functions. Especially with known (non 0) reliability. Thanks for such a useful starting point!
