### 2021-03-15 sdparsons/splithalf#10 @rMassimiliano
Enhance documentation for exported functions 
In the help for the functions `multiverse.plot`, `threshold`, `splithalf.multiverse`,`testretest.multiverse`, the `Description` section is missing.

Also, consider adding a simple working examples in the `Examples` sections of the documentation, or at the very least a link to the specific section of the "online documentation"

In my opinion all exported functions should have a clear documentation.

https://github.com/openjournals/joss-reviews/issues/3041
