### 2021-03-15 sdparsons/splithalf#11 @rMassimiliano
 `speedtestdata` not properly loaded
It looks the data set `speedtestdata` in the package is not properly loaded/codified.

I would also consider using more informative heading for the columns
to replace `V1`,`V2`,`V3`, `V4`,	`V5`,  and explain their meaning in the documentation

```
> class(speedtestdata)
[1] "data.frame"
> dim(speedtestdata)
[1] 225   1

>  head(speedtestdata)
                  X.V1.V2.V3.V4.V5
1  1,20,1,80,500,0.659999999999854
2   2,20,1,80,1000,0.9699999999998
3  3,20,1,80,5000,4.32000000000016
4            4,20,1,80,10000,12.79
5 5,20,1,80,20000,19.3900000000003
6 6,20,1,160,500,0.699999999999818
```

https://github.com/openjournals/joss-reviews/issues/3041
