### 2021-03-15 #3041 @rMassimiliano
Hi @oliviaguest, I also think that the package overall is in good shape. I opened a couple of issues reguarding the documentation but nothing major so far.

Hi @sdparsons I also have some minor suggestions on the `Readme.md`, in addition to the ones proposed by @Nathaniel-Haines

### Comments on the readme:

* In the section **A note on pre-processing**, I think that in the first snippet
the function `filter` is also from the `dplyr` package, so I would write `dplyr::filter`, to be consistent with the rest.

* In the section **splithalf runtime?**, I would add the unit of measure  for the runtime. Maybe on the figure's y axis?

* All the figures look blurry on my screen, I would consider save them at an higher resolution, if possible.
