### 2021-03-20 #3041 @jpsnijder
Hi @sdparsons, I have re-read the documentation and paper and I have no further suggestions or comments.
I think this is a very helpful software based on proper theoretical background knowledge that increases the accuracy of reliability assessment. Thank you for this addition to the measurement field. Great work.
@oliviaguest
