### 2021-03-25 2f8bf56
initial JOSS suggested improvements
