+++ README.md
@@ -36,9 +36,11 @@ proposed that psychological science should set a standard expectation
 for the reporting of reliability information for cognitive and
 behavioural measures (2019). **splithalf** was developed to support this
 proposition by providing a tool to easily extract internal consistency
-reliability estimates from behavioural measures. \#\# Installation
+reliability estimates from behavioural measures.
 
-The latest release version (`0.7.1` unofficial version name: Kitten
+## Installation
+
+The latest release version (`0.7.2` unofficial version name: Kitten
 Mittens) can be installed from CRAN:
 
 ``` r
@@ -57,9 +59,9 @@ functions. The **robustbase** package is used to extract median scores
 when applicable. The computationally heavy tasks (extracting many random
 half samples of the data) are written in `c++` via the `R` package
 **Rcpp** (Eddelbuettel et al. 2018). Figures use the **ggplot** package
-(Wickham 2016), raincloud plots use code adapted from Allen et
-al. (Allen et al. 2019), and the **patchwork** package (Pedersen 2019)
-is used for plotting the multiverse analyses.
+(Wickham 2016), raincloud plots use code adapted from Allen et al.
+(Allen et al. 2019), and the **patchwork** package (Pedersen 2019) is
+used for plotting the multiverse analyses.
 
 ### Citing the package
 
@@ -90,7 +92,7 @@ usability due to people asking for help.
 
 **Now on github and submitted to CRAN: VERSION 0.7.2 \[unofficial
 version name: “Kitten Mittens”\]** Featuring reliability multiverse
-analyses\!\!\!
+analyses!!!
 
 This update includes the addition of reliability multiverse functions. A
 *splithalf* object can be inputted into *splithalf.multiverse* to
@@ -98,7 +100,7 @@ estimate reliability across a user-defined list of data-processing
 specifications (e.g. min/max RTs). Then, because sometimes science is
 more art than science, the multiverse output can be plotted with
 *multiverse.plot*. A brief tutorial can be found below, and a more
-comprehensive one can be found in a recent preprint (Parsons 2020b).
+comprehensive one can be found in a recent preprint (Parsons 2020).
 
 Additionally, the output of splithalf has been reworked. Now a list is
 returned including the specific function calls, the processed data.
@@ -114,17 +116,17 @@ It is important that we have a similar understanding of the terminology
 I use in the package and documentation. Each is also discussed in
 reference to the functions later in the documentation.
 
-  - Trial – whatever happens in this task, e.g. a stimuli is presented.
+-   Trial – whatever happens in this task, e.g. a stimuli is presented.
     Importantly, participants give one response per trial
-  - Trial type – often trials can be split into different trial types
+-   Trial type – often trials can be split into different trial types
     (e.g. to compare congruent and incongruent trials)
-  - Condition - this might be different blocks of trials, or something
+-   Condition - this might be different blocks of trials, or something
     to be assessed separately within the functions. e.g. a task might
     have a block of ‘positive’ trials and a block of ‘negative’ trials.
-  - Datatype - I use this to refer to the outcome of interest.
+-   Datatype - I use this to refer to the outcome of interest.
     specifically whether one is interested in average response times or
     accuracy rates
-  - Score - I use score to indicate how the final outcome is measured;
[…]
