+++ DESCRIPTION
@@ -26,6 +26,6 @@ Suggests:
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.1.0
+RoxygenNote: 7.1.1
 URL: https://github.com/sdparsons/splithalf
 BugReports: https://github.com/sdparsons/splithalf
