### 2021-03-25 sdparsons/splithalf#11 @sdparsons
thanks @rMassimiliano
I've converted the data from .csv to .rda and it now loads correctly

> data(speedtestdata)
> class(speedtestdata)
[1] "data.frame"
> dim(speedtestdata)
[1] 225   6
> head(speedtestdata)
  Sim sample_size Number_conditions trials permutations runtime
1   1          20       1 condition     80          500    0.66
2   2          20       1 condition     80         1000    0.97
3   3          20       1 condition     80         5000    4.32
4   4          20       1 condition     80        10000   12.79
5   5          20       1 condition     80        20000   19.39
6   6          20       1 condition    160          500    0.70
