### 2021-03-25 ba3a18e
fixing speedtest data
