+++ inst/book.bib
@@ -15,6 +15,7 @@
   edition = {2nd},
   note = {ISBN 978-1498716963},
   url = {http://yihui.name/knitr/},
+  doi = {10.1201/b15166}
 }
 
 @Manual{R-bookdown,
