### 2021-03-25 #3041 @sdparsons
Thanks @oliviaguest, music to my ears :)
Hi everyone, thank you so much for the through review!
I have begun revising based on the issues raised, notes on changes below:
- improved the runtime figure (added seconds, improved the legend and facet labels)
- made minor corrections, including correcting the installed version number to 0.7.2
- the speedtestdata.csv now has variable headings and loads correctly (as raised here [sdparsons/splithalf#11](https://github.com/sdparsons/splithalf/issues/11) )
- included dplyr:: and tidyr:: where appropriate
- added missing doi from this message: 10.1201/b15166 may be a valid DOI for title: Dynamic Documents with R and knitr
