### 2021-03-29 #3041 @sdparsons
I do still have the bigger issues that @Nathaniel-Haines and @rMassimiliano raised to fix:
- the plotting issue
- the simulation/validation issue
- enhancing export documentation
But, I think these can be managed in short time and I'll keep y'all updated:)
