### 2021-04-01 c7bd434
working on validation script
