+++ R/splithalf.R
@@ -535,7 +535,7 @@ splithalf <- function(data,
           cor2 = cor(difference2_1, difference2_2, use = "pairwise.complete"),
           n = dplyr::n()
         ) %>%
-        dplyr::mutate(splithalf = round((cor1 + cor2) / 2), round.to) %>%
+        dplyr::mutate(splithalf = round((cor1 + cor2) / 2, round.to)) %>%
         dplyr::mutate(spearmanbrown = round((2 * splithalf) / ((1 + (2 - 1) * abs(splithalf)))), round.to) %>%
         dplyr::select(-1, -2)
 
