+++ vignettes/sim-validation.Rmd
@@ -1,0 +1,736 @@
+---
+title: "sim-validation"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{sim-validation}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+knitr::opts_chunk$set(echo = TRUE)
+knitr::opts_chunk$set(message = FALSE)
+knitr::opts_chunk$set(warning = FALSE)
+
+```
+
+
+```{r setup}
+library("splithalf")
+library("tidyr")
+
+```
+
+# Usage
+
+This script runs _splithalf_ with many possible combinations of the options. Its main use is to ensure that the _splithalf_ function 
+
+We want to run the following:
+
+outcome x score x halftype x average = 3x3x2 = 18
+
+* outcome = "RT"
+* outcome = "accuracy"
+
+* score = "average"
+* score = "difference" (the default option)
+* score = "difference_of_difference"
+
+* halftype = "oddeven"
+* halftype = "halfs"
+* halftype = "random"
+
+* average = "mean"
+* average = "median"
+
+
+```{r generate_data}
+# Thanks to Nathaniel Haines for the suggestion and the starting code (https://github.com/sdparsons/splithalf/issues/9). 
+
+set.seed(01042021)
+
+n_participants <- 60 ## sample size
+n_trials <- 80
+n_blocks <- 2
+
+sim_data <- data.frame(participant_number = rep(1:n_participants, each = n_blocks * n_trials),
+                       trial_number = rep(1:n_trials, times = n_blocks * n_participants),
+                       block_name = rep(c("A","B"), 
+                                        each = n_trials, 
+                                        length.out = n_participants * n_trials * n_blocks),
+                       trial_type = rep(c("congruent","congruent",
+                                          "incongruent","incongruent"),
+                                        length.out = n_participants * n_trials * n_blocks / 2),
+                       RT = rnorm(n_participants * n_trials * n_blocks, 500, 200),
+                       ACC = 1)
+
+sim_data_acc <- data.frame(participant_number = rep(1:n_participants, each = n_blocks * n_trials),
+                       trial_number = rep(1:n_trials, times = n_blocks * n_participants),
+                       block_name = rep(c("A","B"), 
+                                        each = n_trials, 
+                                        length.out = n_participants * n_trials * n_blocks),
+                       trial_type = rep(c("congruent","congruent",
+                                          "incongruent","incongruent"),
+                                        length.out = n_participants * n_trials * n_blocks / 2),
+                       RT = rnorm(n_participants * n_trials * n_blocks, 500, 200),
+                       ACC = c(rbinom(n_participants *
+                                    n_trials *
+                                    n_blocks / 6, 
+                                    1, .5),
+                               rbinom(n_participants *
+                                    n_trials *
+                                    n_blocks / 6, 
+                                    1, .7),
+                               rbinom(n_participants *
+                                    n_trials *
+                                    n_blocks / 6, 
+                                    1, .9),           
+                               rbinom(n_participants *
+                                    n_trials *
+                                    n_blocks / 6, 
+                                    1, .5),
+                               rbinom(n_participants *
+                                    n_trials *
+                                    n_blocks / 6, 
+                                    1, .7),
+                               rbinom(n_participants *
+                                    n_trials *
+                                    n_blocks / 6, 
[…]
