+++ DESCRIPTION
@@ -29,3 +29,4 @@ LazyData: true
 RoxygenNote: 7.1.1
 URL: https://github.com/sdparsons/splithalf
 BugReports: https://github.com/sdparsons/splithalf
+VignetteBuilder: knitr
