### 2021-04-01 sdparsons/splithalf#9 @sdparsons
I have added a validation script to the vignettes (https://github.com/sdparsons/splithalf/tree/master/vignettes). This was a really useful exercise as I was able to catch several minor bugs with some of the combinations.

Currently the main use is to check that the function works, rather than speaking entirely to the validation point - but it's a step in the right direction :)

It currently runs all combinations of the following options: outcome x score x halftype x average = 2x3x3x2 = 36

* outcome = "RT"
* outcome = "accuracy"

* score = "average"
* score = "difference" (the default option)
* score = "difference_of_difference"

* halftype = "oddeven"
* halftype = "halfs"
* halftype = "random"

* average = "mean"
* average = "median"
