### 2021-04-03 sdparsons/splithalf#8 @Nathaniel-Haines
@sdparsons awesome, it works on my end too! I will go ahead and close the issue now 👍
