### 2021-04-03 sdparsons/splithalf#9 @Nathaniel-Haines
I am happy the tests were helpful! I agree on this being a step in the right direction.

Do you plan on adding a parameter recovery type of validation, or do you think these checks are sufficient for now? (I just want to make sure I am not holding this open if we think the current checks are reasonable for now)
