### 2021-04-06 sdparsons/splithalf#9 @sdparsons
What I'd really like to do is figure out how to simulate data with specified reliabilities in order to test parameter recovery. I've not had enough of a chance to look into this for difference scores in particular - and for internal consistency, rather than test retest.
Leave it with me for now :)
