+++ NEWS.Rmd
@@ -13,7 +13,17 @@ knitr::opts_chunk$set(
 ```
 
 
-splithalf v0.7.2 (current development version) [unofficial version name: Kitten Mittens]
+UPCOMING: splithalf v0.8.1 (current development version - work in progress) [unofficial version name: Rum Ham]
+==============
+changes:
+* added validation vignette (tests many combinations of options within splithalf)
+* fixed minor bugs in some combinations of options within splithalf (typically within the )
+* fixed some documentation
+* fixed plotting issue with score = "difference_of_difference"
+* minor improvements to documentation
+
+
+splithalf v0.7.2 [unofficial version name: Kitten Mittens]
 ==============
 changes:
 * minor updates: reworked documentation and README 
@@ -21,7 +31,7 @@ changes:
 * bug fixes to splithalf.multiverse
 
 
-splithalf v0.7.1 (current development version) [unofficial version name: Kitten Mittens]
+splithalf v0.7.1 [unofficial version name: Kitten Mittens]
 ==============
 changes:
 * added multiverse analyses to package. Functions _splithalf.multiverse_, _testretest.multiverse_, _plot.multiverse_, and _threshold_ now available!
