### 2021-04-06 d62e716
adding documentation to functions
