### 2021-04-06 3aa9bfe
adding documentation to functions
