+++ R/splithalf_multiverse.R
@@ -1,5 +1,6 @@
 #' Multiverse of data processing decisions on internal consistency reliability estimates.
 #'
+#' This function enables the user to run a multiverse of data processing options and extract the resulting reliabilty estimates generated by splithalf. The user specifies a set of data processing decisions and passes this to the function, along with a splithalf object. The output can then be explored and plotted as desired.
 #' The (unofficial) function version name is "This function will let you get honey from a hornets nest"
 #' @param input splithalf object or list of splithalf objects
 #' @param specifications list of data processing specifications
