+++ R/plot_multiverse.R
@@ -1,5 +1,6 @@
 #' Visualising reliability multiverses
 #'
+#' This function allows the user to plot the output from splithalf_multiverse or testretest_multiverse. The plot includes an upper panel with all reliabilty estimates (and CIs) and a lower panel that indicates the data processing specifications corresponding to that reliability estimate.
 #' The (unofficial) function version name is "This function will make you a master in bird law"
 #' @param multiverse multiverse object or list of multiverse objects
 #' @param title add a title to the plot?
