+++ NAMESPACE
@@ -9,6 +9,7 @@ import(Rcpp)
 import(ggplot2)
 import(grid)
 import(patchwork)
+import(psych)
 import(tidyr)
 importFrom(Rcpp,sourceCpp)
 importFrom(dplyr,filter)
