### 2021-04-06 sdparsons/splithalf#10 @sdparsons
thanks @rMassimiliano!
As an interim update - I've done this for the functions splithalf, splithalf_multiverse, threshold, and plot_multiverse.
I am working to improve the flexibility of the testretest_multiverse function, and will add the documentation and examples for this asap too
