### 2021-04-09 sdparsons/splithalf#9 @sdparsons
On reflection from the twitter conversation about this, I suspect that a full parameter recovery simulation might be big enough to almost constitute a separate project. So, for now I think that aspect of this issue could be closed.

What I will still do is build a vignette to check that all of the multiverse options and related functions work across input parameters.
