### 2021-04-10 sdparsons/splithalf#9 @Nathaniel-Haines
Yes that is what I was beginning to worry about upon thinking of this further! That said, I think we can consider this closed, and I can wrap up my end of the review 😄
