### 2021-04-10 #3041 @Nathaniel-Haines
@oliviaguest, given both issues I opened are now closed (https://github.com/sdparsons/splithalf/issues/9 and https://github.com/sdparsons/splithalf/issues/8), I am now signed off! I finalized the checklist, too.

Congrats on the package, @sdparsons! 🥳
