### 2021-04-12 #3041 @sdparsons
I've left open @rMassimiliano 's documentation issue open (https://github.com/sdparsons/splithalf/issues/10) as I need to sort this for the testretest.multiverse function. Currently working on it and disliking past Sam for not thinking forward and building in all of the same input checks as the main functions...
