+++ R/testretest_multiverse.R
@@ -1,7 +1,7 @@
 #' Multiverse of data processing decisions on test retest reliability estimates.
 #'
 #' The (unofficial) function version name is "This function will help you pay the troll toll"
-#' @param input list of two datasets
+#' @param data dataset
 #' @param specifications list of data processing specifications
 #' @param test correlation, ICC2, r ICC3
 #' @param var.participant = "subject",
@@ -20,37 +20,94 @@
 #' @import ggplot2
 #' @import grid
 #' @import patchwork
-#' @import psych
 #' @importFrom stats complete.cases cor median na.omit quantile sd cor.test
 #' @importFrom robustbase colMedians
-#' @importFrom dplyr select summarise group_by mutate n_distinct filter ungroup
+#' @importFrom dplyr select summarise group_by mutate n_distinct filter ungroup n
 #' @importFrom tidyr gather
 #' @importFrom plyr arrange
+#' @importFrom psych ICC
 #' @useDynLib splithalf, .registration = TRUE
 #' @importFrom Rcpp sourceCpp
 #' @importFrom utils setTxtProgressBar txtProgressBar capture.output
 #' @export
 
-testretest.multiverse <- function(input,
-                                  specifications,
-                                  test = "ICC2",
-                                  var.participant = "subject",
-                                  var.ACC = "correct",
-                                  var.RT = "RT") {
+testretest.multiverse <- function(data,
+                                   specifications,
+                                   test = "ICC2",
+
+                                   outcome = "RT",
+                                   score = "difference",
+
+                                   var.participant = "subject",
+                                   var.ACC = "correct",
+                                   var.RT = "RT",
+                                   var.time = "time",
+                                   var.compare = "congruency",
+                                   compare1 = "Congruent",
+                                   compare2 = "Incongruent") {
 
 
-  # you'll need something here to differentiate between internal consistency and test-retest.
+  # check that the dataframe is a data frame
+  if (is.data.frame(data) == FALSE) {
+    stop("a data frame has not been specified in data = ")
+  }
+  # check for missing variables
+  if (outcome != "RT" & outcome != "accuracy") {
+    stop("the outcome has not been specified: select from RT or accuracy")
+  }
+  if (score != "average" &
+      score != "difference" &
+      score != "difference_of_difference") {
+    stop(
+      "the score has not been specified: select from average, difference, or difference_of_difference"
+    )
+  }
+  if(score == "DPrime") {
+    warning("the DPrime score is under development. There are many versions of d prime, please check this is the correct version for your analyses")
+  }
 
-  ### new
-  subject = var.participant
-  correct = var.ACC
+  # check that all of the variables exist in the data frame,
+  # including the trial level components
+  if (var.RT %in% colnames(data) == FALSE & outcome != "accuracy") {
+    stop("the RT varible has not been specified")
+  }
+  if (var.participant %in% colnames(data) == FALSE) {
+    stop("the participant varible has not been specified")
+  }
+  if (score == "difference" | score == "difference_of_difference") {
+    if (var.compare %in% colnames(data) == FALSE) {
+      stop("the compare varible has not been specified")
+    }
+    if (compare1 %in% unique(data[[var.compare]]) == FALSE) {
+      stop("compare1 does not exist in the compare variable")
+    }
+    if (compare2 %in% unique(data[[var.compare]]) == FALSE) {
[…]
