+++ vignettes/sim-validation.Rmd
@@ -25,27 +25,11 @@ library("tidyr")
 
 ```
 
-# Usage
+# Validaion Sim for splithalf
 
 This script runs _splithalf_ with many possible combinations of the options. Its main use is to ensure that the _splithalf_ function 
 
-We want to run the following:
-
-outcome x score x halftype x average = 2x3x3x2 = 36
-
-* outcome = "RT"
-* outcome = "accuracy"
-
-* score = "average"
-* score = "difference" (the default option)
-* score = "difference_of_difference"
-
-* halftype = "oddeven"
-* halftype = "halfs"
-* halftype = "random"
-
-* average = "mean"
-* average = "median"
+# Splithalf()
 
 
 ```{r generate_data}
@@ -131,7 +115,7 @@ expand.grid(params) %>%
 
 ```
 
-# Response times
+## Response times
 
 ```{r rt_test1}
 splithalf(data = sim_data,
@@ -427,7 +411,7 @@ splithalf(data = sim_data,
           plot = TRUE)
 ```
 
-# Average accuracy rates
+## Average accuracy rates
 
 ```{r acc_test1}
 splithalf(data = sim_data_acc,
@@ -742,3 +726,162 @@ splithalf(data = sim_data_acc,
           plot = TRUE)
 ```
 
+# multiverse
+
+Runs several multiverse 
+
+```{r generate_mv_data}
+set.seed(01042021)
+
+n_participants <- 80 ## sample size
+n_trials <- 120
+n_blocks <- 2
+
+
+sim_data_mv <- data.frame(participant_number = rep(1:n_participants, 
+                                                    each = n_blocks * n_trials),
+                           trial_number = rep(1:n_trials, 
+                                              times = n_blocks * n_participants),
+                           block_name = rep(c(1,2), 
+                                            each = n_trials, 
+                                            length.out = n_participants * n_trials * n_blocks),
+                           trial_type = rep(c("congruent","congruent",
+                                              "incongruent","incongruent"),
+                                            length.out = n_participants * n_trials * n_blocks / 2),
+                           RT = rnorm(n_participants * n_trials * n_blocks, 
+                                      500, 
+                                      200),
+                           ACC = c(rbinom(n_participants *
+                                            n_trials *
+                                            n_blocks / 6, 
+                                          1, .5),
+                                   rbinom(n_participants *
+                                            n_trials *
+                                            n_blocks / 6, 
+                                          1, .7),
+                                   rbinom(n_participants *
+                                            n_trials *
+                                            n_blocks / 6, 
+                                          1, .9),           
+                                   rbinom(n_participants *
+                                            n_trials *
+                                            n_blocks / 6, 
+                                          1, .5),
+                                   rbinom(n_participants *
+                                            n_trials *
+                                            n_blocks / 6, 
+                                          1, .7),
+                                   rbinom(n_participants *
+                                            n_trials *
+                                            n_blocks / 6, 
+                                          1, .9)))
+
+
+```
+
+
+```{r create_specs}
+specifications <- list(
+  ACC_cutoff = c(0, 0.5),  
+  RT_min           = c(0, 200),
+  RT_max            = c(2000, 3000),
+  RT_sd_cutoff      = c(0, 2),
+  split_by          = c("subject", "trial"), 
+  averaging_method  = c("mean")
+)
+
+```
+
+## internal consistency
+
+```{r splithalf_for_mv}
+mv_splithalf <- splithalf(data = sim_data,
+                        outcome = "RT",
+                        score = "difference",
+                        conditionlist = c("A"),
+                        halftype = "random",
+                        permutations = 5000,
+                        var.RT = "RT",
+                        var.condition = "block_name",
+                        var.participant = "participant_number",
[…]
