+++ man/splithalf.multiverse.Rd
@@ -15,6 +15,7 @@ splithalf.multiverse(input, specifications)
 Returns a multiverse object containing the reliability estimates and dataframes from all data processing specifications provided
 }
 \description{
+This function enables the user to run a multiverse of data processing options and extract the resulting reliabilty estimates generated by splithalf. The user specifies a set of data processing decisions and passes this to the function, along with a splithalf object. The output can then be explored and plotted as desired.
 The (unofficial) function version name is "This function will let you get honey from a hornets nest"
 }
 \examples{
