+++ NAMESPACE
@@ -8,6 +8,7 @@ export(threshold)
 import(Rcpp)
 import(ggplot2)
 import(grid)
+import(lme4)
 import(patchwork)
 import(tidyr)
 importFrom(Rcpp,sourceCpp)
