### 2021-04-12 sdparsons/splithalf#10 @sdparsons
Hey @rMassimiliano I have now added documentation help for the testretest.multiverse function and included the multiverse functions in the validation sim in the vignettes.
