### 2021-04-12 #3041 @sdparsons
I have now sorted this for the testretest.multiverse function. I believe that this ticks off the last of the major issues.

Given that combined these were some bigger changes, I plan to update the package version to 0.8.1 (and relevant documentation) and resubmit to CRAN - assuming that there are no further outstanding issues to fix?
