### 2021-04-14 #3041 @rMassimiliano
Hi @sdparsons, good job! A couple of minor suggestions for the documentations in the following. I don't think they require an issue.

1. In the documentation there are several sentences like `The (unofficial) function version name is  "This function will make you a master in bird law"` is this on purpose? In case fine by me. Maybe I just don't get the jokes.

2. I do not get why  `multiverse.plot()` and  `threshold()` documentations are lumped together. I find this a bit hard to follow. If there is a good reason fine by me.

3. When you have character string arguments It is a good practice to specify the values that the arguments are allowed to take.
For example in the function  `splithalf()`

* `average: use mean or median to calculate average scores?`

 could be something like:` a character string specifying the function to calculate the average scores, must be one of  ‘"mean"’ or ‘"median"’`. If it is an appropriate description.  It will make your software easier to use for the final users.

4. a couple of small typos in the function  `splithalf()`

*  `var.ACC: specifiec the accuracy variable name in data`
    I guess `specifiec` should be `specifies`

* in `var.compare: specified the variable that is used to calculate
          difference scores (e.g. including congruent and incongruent
          trials)`

I would change `specified` into `specifies` as for the other arguments.

*  `compare2: specifies the first trial type to be compared (e.g.
          incongruent trials)`
     maybe the second trial?

5. `testretest.multiverse()`

Description:

     The (unofficial) function version name is "This function will help
     you pay the troll toll"

maybe I will add some lines on why the function is useful.
