### 2021-04-14 d646a81
improvinf documentation and fixing typos
