+++ R/splithalf_multiverse.R
@@ -1,6 +1,7 @@
 #' Multiverse of data processing decisions on internal consistency reliability estimates.
 #'
-#' This function enables the user to run a multiverse of data processing options and extract the resulting reliabilty estimates generated by splithalf. The user specifies a set of data processing decisions and passes this to the function, along with a splithalf object. The output can then be explored and plotted as desired.
+#' This function enables the user to run a multiverse of data processing options and extract the resulting (internal consistency) reliability estimates generated by splithalf. The user specifies a set of data processing decisions and passes this to the function, along with a splithalf object. The output can then be explored and plotted as desired.
+#'
 #' The (unofficial) function version name is "This function will let you get honey from a hornets nest"
 #' @param input splithalf object or list of splithalf objects
 #' @param specifications list of data processing specifications
@@ -57,14 +58,6 @@
 #' ## can be plot with:
 #' multiverse.plot(multiverse = multiverse,
 #'                 title = "README multiverse")
-#'
-#' ## the threshold function can be used to return the number of estimates
-#' ## above or below a certain threshold
-#'
-#' threshold(multiverse = multiverse,
-#'           threshold = 0.7,
-#'           use = "estimate",
-#'           dir = "above")
 #' }
 #' @import tidyr
 #' @import Rcpp
