+++ R/zzz.R
@@ -7,7 +7,7 @@
                       fields="Version")
   packageStartupMessage("This is ",paste(pkgname, version))
   packageStartupMessage(pkgname, " is BETA software! Please report any bugs.")
-  packageStartupMessage("The (unofficial) version name is: Kitten Mittens")
+  packageStartupMessage("The (unofficial) version name is: Rum Ham")
   packageStartupMessage("For documentation, questions, and issues, please see github.com/sdparsons/splithalf")
   packageStartupMessage("or email sam.parsons@psy.ox.ac.uk")
 }
