### 2021-04-14 #3041 @sdparsons
thanks @rMassimiliano
1. Yeah, these are little easter-egg references to It's always sunny in Philadelphia. Largely for my own amusement.
2. Fair point. Originally it was because the set-up code is needed anyway, so I didnt think it would confuse things to include this too. But, I've removed the example code that is not relevant to that function.
3. Good point, I've worked through to catch these.
4. corrected :)
5. description added

I took this opportunity to update the version to 0.8.1 to capture all of the updates and fixes that arose from the review - thank you all!
Also ran through devtools::check() and everything passed :)
