### 2021-04-14 #3041 @jpsnijder
Please disregard that last email, it was for Sam.

My apologies,

Jean-Paul

Sent from my iPhone

> On Apr 14, 2021, at 22:55, Jean-paul Snijder ***@***.***> wrote:
>
> Honestly, I would love to go toe to toe with you in bird law... filibuster..
