### 2021-04-15 #3041 @oliviaguest
@jpsnijder no worries — deleted them from this public thread. 🌼

@rMassimiliano I think only "Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?" is left now, right?

@sdparsons this is looking very close to ready: exciting! ☺️
