### 2021-04-15 #3041 @rMassimiliano
Hi @oliviaguest all box checked. The package looks fine to me 😄

@sdparsons great job! The documentation looks much nicer. Here three typos I spotted:

1. `multiverse.plot()`

* in the description     `reliabilty` -> `reliabiltiy`

* in value    `visualistation` -> `visualization`

2. `splithalf()`

* in the description `ourcomes` -> `outcomes`
