### 2021-04-15 #3041 @sdparsons
This process is definitely highlighting how bad I am at proofing text in RStudio...
Thanks for catching these @rMassimiliano - all fixed now. :)
