+++ paper.md
@@ -15,8 +15,8 @@ affiliations:
    index: 1
 
 citation_author: Parsons
-date: 30 December 2020
-year: 2020
+date: 19 April 2021
+year: 2021
 bibliography: inst/book.bib
 ---
 
@@ -24,7 +24,7 @@ bibliography: inst/book.bib
 
 The `R` package **splithalf** provides tools to estimate the internal consistency reliability of cognitive measures. In particular, the tools were developed for application to tasks that use difference scores as the main outcome measure, for instance the Stroop score or dot-probe attention bias index (average RT in incongruent trials minus average RT in congruent trials). 
 
-The methods in **splithalf** are built around split half reliability estimation. To increase the robustness of these estimates, the package implements a permutation approach that takes a large number of random (without replacement) split halves of the data. For each permutation the correlation between halves is calculated, with the Spearman-Brown correction applied [@spearman_proof_1904]. This process generates a distribution of reliability estimates from which we can extract summary statistics (e.g. average and 95% HDI).
+The methods in **splithalf** are built around split half reliability estimation. To increase the robustness of these estimates, the package implements a permutation approach that takes a large number of random (without replacement) split halves of the data. For each permutation the correlation between halves is calculated, with the Spearman-Brown correction applied [@spearman_proof_1904]. This process generates a distribution of reliability estimates from which we can extract and plot summary statistics (e.g. average and 95% HDI).
 
 ### Statement of need
 
@@ -34,19 +34,19 @@ While many cognitive tasks yield highly robust effects, the same task may not yi
 
 The _splithalf_ function estimates reliability for a variety of scoring types, including average scores, difference scores, and difference-of-difference scores. The function is also extendable to other scoring calculations, e.g. signal detection. Reliability can be estimated for response time and accuracy rate outcomes. A plotting option is also provided for users to examine the distribution of reliability estimates. 
 
-Users can perform a reliability multiverse analysis to examine the stability of reliability estimates, given a rage of data processing decisions [@parsons_exploring_2020]. By passing a list of data processing decisions (e.g. maximum and minimum RT cutoffs) and a _splithalf_ object into the _splithalf.multiverse_ function. The output can then be plotted with _plot_multiverse_. 
+Users can perform a reliability multiverse analysis to examine the stability of reliability estimates, given a rage of data processing decisions [@parsons_exploring_2020]. By passing a list of data processing decisions (e.g. maximum and minimum RT cutoffs) and a _splithalf_ object into the _splithalf.multiverse_ function. _testretest.multiverse_ can be used to explore a similar multiverse of data processing decisions on test retest reliability. Multiverse output can be plotted with _plot_multiverse_. 
 
 
-**splithalf** requires the **tidyr** [@R-tidyr] and **dplyr** [@R-dplyr] packages for data handling within the functions. The **robustbase** package is used to extract median scores when applicable. The computationally heavy tasks (extracting many random half samples of the data) are written in `c++` via the `R` package **Rcpp** [@R-Rcpp]. Figures use the **ggplot** package [@R-ggplot2], raincloud plots use code adapted from Allen et al. [@allen_raincloud_2019], and the **patchwork** package [@R-patchwork] is used for plotting the multiverse analyses. 
[…]
