### 2021-04-23 #3041 @sdparsons
This is very exciting, thank you all!
A huge thank you to @jpsnijder, @Nathaniel-Haines, @rMassimiliano. Your input pushed me to fix a bunch of issues I would not have spotted if it were not for your input and massively improve the documentation around splithalf. This has by far felt like the most productive review process I have been part of. Thank you @oliviaguest too for editing and guiding all of us through this great process. I really appreciate all of your time and expertise. I've added my name to the reviewer list, so hopefully I'll be on the other side of this process soon to give back :)
